feat：拆分关系构建和关系信息提取
